Add Post representation using JPA
JPA Entities are POJOs associated to data that can be persisted to
database. `Post` Entity represents a table of posts in the DB. Therefore
an instance of a `Post` can be used to manage a row in this table.

PostRepository it's an interface that extends JPARepository. By doing that, it
inherits different methods that can be used for common persistence operations.